Fix for unsigned ID
Some databases (namely PostGres and MariaDB) require the types of foreign and linked primary keys to be the same type. If one is `unsigned` and the other isn't, then the database will throw an issue (see #30). This sets the ID to `unsigned` which is required as per the [Laravel 5 documentation] (https://laravel.com/docs/5.0/schema#foreign-keys).